refactor: observeLoadMore 코드 feature 폴더로 이동
